pyproject.toml
[project]
name = "recommandation-engine"
version = "0.1.0"
description = "Moteur de recommandation de produits basé sur ratings et analyse de sentiment"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.4.0",
    "textblob>=0.18.0",
    "openpyxl>=3.1.0",
    "faker>=40.4.0",
    "requests>=2.32.5",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]
